+++ game/components/system_list.h
@@ -19,12 +19,13 @@
 #include "s_timer.h"
 #include "s_patrol.h"
 #include "s_scene_transition.h"
+#include "s_noise.h"
 
 //Extra
-#include "s_fire.h"
 #include "s_player.h"
 
 #define PHYSICS_UPDATE_SYSTEMS System::Actor::update(c); \
+                               System::Noise::update(c); \
                                System::Camera::update(c);
 
 #define RENDER_SYSTEMS System::Camera::prerender(c); \
@@ -32,6 +33,6 @@
                        if (not c.tme_active) { \
                            System::Texture::render(c); \
                            System::Text::render(c); \
-                           System::Fire::render(c); \
+                           System::Noise::render(c); \
                        }